# openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Fonoteca.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/files": {
      "get": {
        "tags": [
          "Files"
        ],
        "summary": "One directory of the library, from disk, annotated from the catalogue.",
        "description": "Lists what is actually on disk — including files nothing has scanned and files that are not audio at all — and puts the catalogue's own counts on each row: how many audio files it holds beneath, how many are identified, and how many are filed under an album.",
        "operationId": "ListLibraryFolder",
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FolderListing"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/files/content": {
      "get": {
        "tags": [
          "Files"
        ],
        "summary": "The bytes of one file, with range support.",
        "description": "What an <img>, an <audio> or a partial fetch of a rip log reads. Range requests are handled, so seeking into a large FLAC does not read the part before it.\n\n**The media type comes from an allowlist, never from the extension itself.** Anything not on it is served as application/octet-stream and as an attachment, and every response carries X-Content-Type-Options: nosniff — an .html in a music folder is a download, not a page on this origin.",
        "operationId": "GetLibraryFileContent",
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/files/detail": {
      "get": {
        "tags": [
          "Files"
        ],
        "summary": "What one file says about itself, by path.",
        "description": "The decoder's measurements and the file's own tags, for any audio file in the library — including one no scan has seen yet, which is why this is keyed on a path rather than on a media file id.\n\n**It cannot fail on the file.** A truncated FLAC, an unmounted volume and a tag parser that dereferences null all come back as a reading with a note in it, because a screen that 500s on the file you opened it for is worse than one that says the decoder would not read it.",
        "operationId": "GetLibraryFileDetail",
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FilePreviewResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/files/art": {
      "get": {
        "tags": [
          "Files"
        ],
        "summary": "A picture of one entry: a file's embedded cover, or a folder's own.",
        "description": "For an audio file, the front cover stored inside it, served unchanged. For a directory, the first of cover/folder/front/album it holds.\n\n**404 is the ordinary answer, not an error.** Plenty of files carry no picture and plenty of folders hold none, and the screen draws a monogram for both.",
        "operationId": "GetLibraryEntryArt",
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/catalogue/matching/files/{id}": {
      "get": {
        "tags": [
          "Catalogue"
        ],
        "summary": "Everything one file can be told to say about itself.",
        "description": "What a person needs in front of them while deciding which recording a file holds, and the half the worklist could not carry. Half of it is catalogue — path, size, the length `fpcalc` measured, what each pass concluded — and half of it is read from the bytes on demand: codec, bitrate, sample rate, bit depth, channels, and whatever the file's own tags still claim.\n\n**Separate from the candidate set on purpose.** This costs one header read and no provider request at all, so it answers for every file on the worklist — including the ones whose candidates cannot be recovered, which are most of them. The measured properties are written back to the catalogue as a side effect, since nothing else populates them yet.\n\nNothing here fails the request. An unmounted volume, a moved file or a decoder that refuses the container all come back as an absent `audio` block and a `note` saying which, because a file nothing could read is itself worth knowing about on this screen.",
        "operationId": "GetMatchingFile",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubjectFileResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/catalogue/matching/files/{id}/art": {
      "get": {
        "tags": [
          "Catalogue"
        ],
        "summary": "The cover art embedded in one file, as it is stored.",
        "description": "The bytes of the file's own front cover, served unchanged — no scaling, no re-encoding. It is the one claim about an album that a person can check against a candidate at a glance, and unlike the folder name it travels inside the file.\n\n**404 is the ordinary answer, not an error.** A file with no picture, a file no tag parser will open and an unmounted volume all come back the same way, because the screen draws the same monogram for all three.",
        "operationId": "GetMatchingFileArtwork",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/catalogue/matching/folders/files": {
      "get": {
        "tags": [
          "Catalogue"
        ],
        "summary": "Every file in one folder, and what each of them is matched to.",
        "description": "The whole folder, not the open questions in it — which is the one view of a rip nothing else here offers. The worklist lists what the passes refused, so a folder of sixteen files that were matched wrongly, minus the three that were refused, appears on it as a three-file album; the thirteen that are wrong are on no screen at all.\n\nEach row says whether the file is still an open question and, when it is not, the album, disc, position and track it was filed under, and how certain that was. Ordered by path, so the discs of a set come back in the order they sit in.\n\nA catalogue read and nothing more: no provider call, no file opened, nothing written.",
        "operationId": "GetFolderContents",
        "parameters": [
          {
            "name": "folder",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FolderContentsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AudioQualityRow": {
        "required": [
          "codec",
          "bitrateKbps",
          "sampleRateHz",
          "bitDepth",
          "channels",
          "lossless",
          "tier"
        ],
        "type": "object",
        "properties": {
          "codec": {
            "type": "string"
          },
          "bitrateKbps": {
            "type": "integer",
            "format": "int32"
          },
          "sampleRateHz": {
            "type": "integer",
            "format": "int32"
          },
          "bitDepth": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "channels": {
            "type": "integer",
            "format": "int32"
          },
          "lossless": {
            "type": "boolean"
          },
          "tier": {
            "type": "string"
          }
        }
      },
...
      "FilePreviewResponse": {
        "required": [
          "path",
          "name",
          "sizeBytes",
          "modifiedUtc",
          "kind",
          "mediaType",
          "isAudio",
          "durationSeconds",
          "decodedCleanly",
          "audio",
          "tags",
          "note"
        ],
        "type": "object",
        "properties": {
          "path": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "sizeBytes": {
            "type": "integer",
            "format": "int64"
          },
          "modifiedUtc": {
            "type": "string",
            "format": "date-time"
          },
          "kind": {
            "type": "string"
          },
          "mediaType": {
            "type": "string"
          },
          "isAudio": {
            "type": "boolean"
          },
          "durationSeconds": {
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          },
          "decodedCleanly": {
            "type": [
              "null",
              "boolean"
            ]
          },
          "audio": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/AudioQualityRow"
              }
            ]
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FileTagRow"
            }
          },
          "note": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "FileTagRow": {
        "required": [
          "name",
          "value"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "value": {
            "type": "string"
          }
        }
      },
...
      "FolderContentsResponse": {
        "required": [
          "folder",
          "files",
          "open",
          "items"
        ],
        "type": "object",
        "properties": {
          "folder": {
            "type": "string"
          },
          "files": {
            "type": "integer",
            "format": "int32"
          },
          "open": {
            "type": "integer",
            "format": "int32"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FolderFileRow"
            }
          }
        }
      },
...
      "FolderEntry": {
        "required": [
          "name",
          "path",
          "isDirectory",
          "sizeBytes",
          "modifiedUtc",
          "isAudio",
          "kind",
          "cataloguedFiles",
          "identified",
          "attributed"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "isDirectory": {
            "type": "boolean"
          },
          "sizeBytes": {
            "type": "integer",
            "format": "int64"
          },
          "modifiedUtc": {
            "type": "string",
            "format": "date-time"
          },
          "isAudio": {
            "type": "boolean"
          },
          "kind": {
            "type": "string"
          },
          "cataloguedFiles": {
            "type": "integer",
            "format": "int32"
          },
          "identified": {
            "type": "integer",
            "format": "int32"
          },
          "attributed": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "FolderListing": {
        "required": [
          "path",
          "parent",
          "exists",
          "entries"
        ],
        "type": "object",
        "properties": {
          "path": {
            "type": "string"
          },
          "parent": {
            "type": [
              "null",
              "string"
            ]
          },
          "exists": {
            "type": "boolean"
          },
          "entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FolderEntry"
            }
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "SubjectFileResponse": {
        "required": [
          "mediaFileId",
          "path",
          "folder",
          "name",
          "format",
          "sizeBytes",
          "size",
          "measured",
          "duration",
          "audio",
          "identification",
          "enrichment",
          "attribution",
          "acoustId",
          "tagged",
          "fingerprinted",
          "integrity",
          "lastScannedUtc",
          "recording",
          "release",
          "tags",
          "note"
        ],
        "type": "object",
        "properties": {
          "mediaFileId": {
            "type": "string",
            "format": "uuid"
          },
          "path": {
            "type": "string"
          },
          "folder": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "format": {
            "type": "string"
          },
          "sizeBytes": {
            "type": "integer",
            "format": "int64"
          },
          "size": {
            "type": "string"
          },
          "measured": {
            "type": [
              "null",
              "string"
            ]
          },
          "duration": {
            "type": [
              "null",
              "string"
            ]
          },
          "audio": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/AudioQualityRow"
              }
            ]
          },
          "identification": {
            "type": "string"
          },
          "enrichment": {
            "type": "string"
          },
          "attribution": {
            "type": "string"
          },
          "acoustId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "tagged": {
            "type": "boolean"
          },
          "fingerprinted": {
            "type": "boolean"
          },
          "integrity": {
            "type": "string"
          },
          "lastScannedUtc": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "recording": {
            "type": [
              "null",
              "string"
            ]
          },
          "release": {
            "type": [
              "null",
              "string"
            ]
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FileTagRow"
            }
          },
          "note": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
    }
  },
...
}